Don't wipe out borderStyle when side border styles are set to null
This adds a new util function for calculating only those border styles
for sides which have been passed in props. For example, if borderLeft
is not set, borderStyleLeft in Block will be undefined, not null.

This is a problem because `null` seems to translate to "none" in React,
wiping out all borders if border is set to `true` and side borders are
all unset.